[InferTypes] Add IntTypeOp, handle inference of concrete types
Add the `hir.int_type` op to construct an unbounded integer type. Extend
the InferTypes pass to unify `hir.inferrable_type` ops with concrete
`hir.int_type` ops. This only works if the inferrable type appears after
the concrete type.

Later on we may want to use a dominance analysis to also allow inference
if the concrete type appears before the inferrable type, but dominates
all users of the inferrable type. Additionally, we may want to move the
concrete type around in the IR to a point where it dominates the
inferrable type.