[InferTypes] Add RefTypeOp, unify concrete types (#5)
Add the `hir.ref_type` op to construct a reference type. The op takes
the type the reference points to as an operand. Extend the InferTypes
pass to unify operations that are equivalent except for their operands.
In this case, create new `hir.unify_type` ops, one for each of the
operands. This allows us to unify for example two reference types, which
will replace one of them with the other and instead create a unification
op for the type inside the reference. This is the first step towards
proper type inference.