[HIR] Add integer constants of arbitrary bit width
Add the `IntAttr` attribute which wraps a `DynamicAPInt`, and integer
with arbitrary bit width. Operations on this type never overflow or see
any other effects of operating on a limited bit width.

Also add the `!hir.int` type and `hir.constant_int` op to materialize
such integers in the IR.