[HIR] Generalize type inference to also support values (#8)
Rename the `hir.inferrable_type` and `hir.unify_type` ops to
`hir.inferrable` and `hir.unify`, respectively. Adjust their definition
to allow them to operate on any type of SSA value, not just `!hir.type`.
This now allows the InferTypes pass to also infer integer constants.
In the future we may want to rename this pass to something more generic.