[HIR] Add EvalConsts pass and constant branches (#9)
Add the EvalConsts pass which fully unrolls any constant control flow
operations in the IR. Also add a `hir.const_br` and `hir.const_cond_br`
operation to encode constant control flow. These currently operate on
regular i1 conditions, which may not be constant. In the future, we'll
want to accept a specific `!hir.const<i1>` type wrapper to enfroce that
the conditions are known at compile time.

Also add a few tests, including a loop test which shows that constant
control flow can be used to implement a static for loop similar to the
following:

    fn foo(a: uint<42>) -> uint<42> {
      let x = a;
      const for i in 0..8 {
        x = process(x);
      }
      return x;
    }