[HIR] Add const type and ops to pack/unpack const values
Add a `!hir.const<T>` wrapper type that can be used to indicate that a
value is a compile-time constant. Also add `hir.const_wrap` and
`hir.const_unwrap` operations to add or remove a layer of `!hir.const`.

In the future we might want to remove the `hir.const_wrap` op again,
since it should not really be possible to conjure up a const value from
a non-const value. This is here for symmetry to get started.