[Syntax] Add AST visitor and printer
Add the necessary infrastructure to systematically traverse the AST with
a visitor and visit each node. Also add an AST printer that uses the
visitor to print the AST in human-readable form to the output. Add a new
`--test-parser` option that makes use of the printer. Finally, add a
test that uses the printer to verify associativity and precedence rules
for expressions.