[Codegen] Try generating different regions of constness
Instead of using `hir.const` ops to wrap const regions of code, create
separate groups of blocks that correspond to different execution phases,
with the deepest nesting of const (highest level of constness) appearing
first in the region.

It's not quite clear yet how this interacts with function arguments.
Should these use a similar block-split approach? Or should we group
function arguments by level of constness, such that we get groups of
arguments passed to the function at different phases of execution?

Unclear.