[HIR] Add callee type operand to CallOp
Add a dedicated operand to `hir.call` that specifies the type of the
called function. We may want to move this onto the callee itself later,
but this is a good starting point. This allows the MIR lowering to know
the type signature of the called function, such that we can create a
properly typed `mir.call` operation.