[HIR] Add unchecked functions, calls, and const regions
Add new HIR operations to represent parts of the IR before their level
of constness has been inferred by a later pass. With these ops we can
add enough information about the constness requirements of values into
the IR to then be able to determine the constness of every IR op in an
analysis.