[Codegen] First steps towards emitting unchecked functions
Use the new `hir.unchecked_func` when generating IR for AST functions.
This will later allow us to do constness inference in a separate
analysis pass, which is a lot cleaner than doing it during codegen. It
is not yet clear how the constness of the arguments themselves should be
represented. At the moment, only the minimum level of constness for each
argument is preserved, by virtue of being used in the type expression of
another argument. This will have strange results, and arguments may need
to be annotated with their constness explicitly.