[HIR] Rename UncheckedConstOp to ExprOp
This op is useful to group the operations that resulted from a single
expression in the AST. If that expression contains side-effects, we want
to propagate constness to all ops involved in the expression, even if
they don't directly contribute to the result value through an SSA edge.

For example:
```
foo({
  sideEffectingCall();
  42
})
```
If `foo` requires its argument to be const, the side effecting call must
also be executed at that constness level, even though it does not
contribute to the result value.