[HIR] Add basic CheckCalls pass
Add a first implementation of a pass that copies function signatures to
the function's call sites. This effectively computes the function
argument and return types at the call site based on the call arguments,
and allows us to insert type unification operations to impose the same
constraints on the call arguments as the function signature does on the
function arguments.

If a call argument has an uninferred type, this pass introduces the
necessary unification ops that allows the argument of the type to be
inferred.

The pass should eventually replace `hir.unchecked_call` ops with some
equivalent op that expresses the fact that its signature has been
checked and detailed type information is available at the call site.
This is not implemented yet.