[HIR] Add first draft of SplitPhases pass
Add a basic pass that splits `hir.unchecked_func` ops into a separate
function per constness level. This will allow us to take a unified
function description as written by the user and split it apart into
distinct phases of compile-time execution. The results of an earlier
phase will then feed into the next phase a constant function arguments
that can be used to monomorphize the later phase.

The pass also performs some basic phase inference for side-effect free
operations. This feature needs to be extended in the future. Also, there
is no error reporting back to the user if an operation uses a value that
will not be defined until a later phase.